Nested grep and writer under the text namespace
Why?
- The functionality in these 2 classes is directly linked to the text edit functionality.
- This will make it easier to add file and folder editing in the future.
